Move verifyCredential method to VerificationRepository
#82
