+++ domain/src/main/java/by/alexandr7035/affinidi_id/domain/repository/CredentialsRepository.kt
@@ -25,8 +25,6 @@ interface CredentialsRepository {
 
     suspend fun getCredentialById(getCredentialByIdReqModel: GetCredentialByIdReqModel, authState: AuthStateModel): Flow<GetCredentialByIdResModel>
 
-    suspend fun verifyCredential(verifyVcReqModel: VerifyVcReqModel): VerifyVcResModel
-
     suspend fun shareCredential(shareVcReq: ShareCredentialReqModel, authState: AuthStateModel): ShareCredentialResModel
 
     suspend fun checkIfHaveCredentialInCache(checkIfHaveVcReqModel: CheckIfHaveVcReqModel): CheckIfHaveVcResModel
